requirements.txt
click==8.1.3
commonmark==0.9.1
hurry==1.1
hurry.filesize==0.9
keystone-engine==0.9.2
prompt-toolkit==3.0.32
Pygments==2.13.0
rich==12.6.0
typer==0.7.0
unicorn==2.0.1
wcwidth==0.2.5
